+++ mingling_core/src/program/repl_exec.rs
@@ -129,9 +129,7 @@ where
                     payload: panic_info,
                 };
                 let program = crate::program::THIS_PROGRAM
-                    .get()
-                    .unwrap()
-                    .as_ref()
+                    .get_raw()
                     .unwrap()
                     .downcast_ref::<Program<C>>()
                     .unwrap();